PY01_DockerRestAPI/requirements.txt
Flask==3.0.3
psycopg2==2.9.9
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2 
python-decouple==3.8
cffi==1.16.0
cryptography==42.0.5
pycparser==2.22
pytz==2024.1
pymongo==4.2.0
dnspython==2.6.1
PyJWT==2.8.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
requests==2.31.0
urllib3==2.2.1
pytest-mock==3.14.0
redis==5.0.3